backend/requirements.txt
flask
flask-socketio
pillow